[FM-6] add preview to HistoryDetailsSummary and move enhancements history to separate composable
