+++ app/src/main/java/com/fish/monsters/features/history/ui/components/HistoryDetailsSummary.kt
@@ -23,10 +23,17 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.fish.monsters.R
+import com.fish.monsters.common.views.PreviewContainer
 import com.fish.monsters.core.database.entities.Contest
+import com.fish.monsters.core.database.entities.contest.Award
 import com.fish.monsters.core.database.entities.contest.AwardsCount
+import com.fish.monsters.core.database.entities.contest.DifficultyLevel
+import com.fish.monsters.core.database.entities.contest.Duration
+import com.fish.monsters.core.database.entities.contest.Enhancement
+import com.fish.monsters.core.database.entities.contest.GameLocation
 import com.fish.monsters.core.theme.DarkPrimaryColor
 
 @Composable
@@ -238,5 +245,37 @@ fun AwardCol(awardCount: AwardsCount) {
             contentDescription = "Award Icon",
             modifier = Modifier.width(17.dp)
         )
+    }
+}
+
+@Preview
+@Composable
+fun HistoryDetailsSummaryPreview() {
+    PreviewContainer {
+        HistoryDetailsSummary(contestDetails = Contest(
+            id = 4,
+            date = "10 march 2024",
+            duration = Duration(1, 15, 40),
+            points = 75,
+            difficultyLevel = DifficultyLevel.MEDIUM,
+            rewardsCount = 12,
+            enhancementsUsed = listOf(
+                Enhancement(
+                    name = "Enhancement 4",
+                    time = Duration(1, 30, 12)
+                )
+            ),
+            bypassedMonsters = 18,
+            awardsEarned = listOf(
+                AwardsCount(Award.Grass, 2),
+                AwardsCount(Award.Flower, 12),
+                AwardsCount(Award.Pumpkin, 1)
+            ),
+            isGameSuccess = true,
+            gameLocation = GameLocation(
+                latitude = 25.0,
+                longitude = 20.0
+            )
+        ))
     }
 }
